Move calendar logic to its own module.
