+++ package.json
@@ -4,6 +4,7 @@
   "description": "Used by french employees to generate their holiday form",
   "main": "index.js",
   "scripts": {
+    "lint": "./node_modules/.bin/eslint src lib",
     "test": "./node_modules/.bin/mocha"
   },
   "repository": {
@@ -21,6 +22,7 @@
   "homepage": "https://github.com/julienw/forms-extension#readme",
   "devDependencies": {
     "chai": "^3.5.0",
+    "eslint": "^3.13.1",
     "jpm": "1.0.1",
     "mocha": "^3.2.0",
     "semver": "5.0.1"
