+++ src/.eslintrc
@@ -1,0 +1,10 @@
+{
+  "extends": "eslint:recommended",
+  "rules": {
+  },
+  "env": {
+    "browser": true,
+    "es6": true,
+    "node": true
+  }
+}
