+++ src/data/results.html
@@ -9,6 +9,7 @@
     <script defer src='async_storage.js'></script>
     <script defer src='fake-data.js'></script>
     <script defer src='french-bank-holidays.js'></script>
+    <script defer src='calendar.js'></script>
     <script defer src='form-persistence.js'></script>
     <script defer src='results-display.js'></script>
     <script defer src='form-interaction.js'></script>
